count up and down

+++ selfdrive/ui/sunnypilot/qt/onroad/hud.h
@@ -73,6 +73,6 @@ private:
   float speedLimitAhead;
   float speedLimitAheadDistance;
   float speedLimitAheadDistancePrev;
-  float speedLimitAheadValidFrame;
+  int speedLimitAheadValidFrame;
   QString roadName;
 };
